:tada: init: add initial project structure using Vite and React
This commit adds the initial structure of the project using Vite as the build tool and React as the rendering library. The necessary dependencies were installed, the src folder was created and an initial component was added to test the configuration. This commit is the base for the future development of the project.