:heavy_plus_sign: build: adding Styled Components dependency
This commit adds the Styled Components dependency to the project, which will allow creating styles in a simpler and more modularized way. Styled Components is a popular library in front-end development that uses CSS syntax in conjunction with JavaScript logic to efficiently create styled components. With this addition, we will be able to improve the appearance and usability of the application in an easier and more organized way.