:fire: removing unnecessary starter files
This commit removes initial files that are no longer needed for the project to continue. During the creation of the project, several files were created that are no longer used, and that only occupy unnecessary space in the project. With this removal, we can keep the code cleaner and more organized, avoiding confusion and possible errors. In addition, this action can contribute to better project performance, since there will be fewer files to be loaded during execution.