:ok_hand: style: removing references to deleted files
This commit removes references to files deleted in the previous commit. As these files no longer exist, it is important that all references to them are removed to avoid possible compilation or execution errors. With this action, we keep the project organized and free of problems that may arise due to unnecessary files or references.