:lipstick: feat: favicon update in app
This commit removes the old favicon from the application and adds a new favicon. The new favicon was created to improve the appearance and visibility of the application. Additionally, a reference to the new favicon has been added to the index.html file to ensure that the new icon displays correctly in the browser's address bar and in bookmarks. With this update, users will have an improved visual experience when using the application.