:heavy_plus_sign: build: added react-icons library
This update adds the react-icons library to the project. This library offers a wide variety of icons that can be used in React components. With the addition of this library, we now have access to high quality icons for use in our project.

Furthermore, the library is easy to integrate with React and offers options for customizing the icons, such as size, color and style. This update was necessary to improve the usability and aesthetics of our application.