:sparkles: feat: added initial structure of Header
This update adds the initial Header structure to the project. A Header component was created that contains the project's logo and a basic navigation menu. The aim is to provide intuitive and easy-to-use navigation for users.

The current structure is basic, but will be improved over time to meet the needs of the project.

This update is important to lay a solid foundation for the project's UI and improve the user experience.