:lipstick: feat: added initial styling to Header component
This update adds initial styling to the Header component. Basic styles have been applied to the logo and navigation menu to improve appearance and usability.

The Header's background color was set to match the project's color palette.

Navigation menu items have been styled with a background color when the user hovers over them to highlight the selected option. Additionally, a hover has been added to the logo to signal that it is clickable.

This update is important to improve the appearance of the Header and make it more attractive and functional for the user. Styling will be improved over time to meet the needs of the project and maintain the visual cohesion of the interface.