:sparkles: feat: added Footer component
This update adds the Footer component to the project. This component is responsible for displaying the gray logo of the project and the copyright phrase.

The Footer component is created with React and styled with CSS. It contains the project's logo in gray and the phrase 'All rights reserved'.

This component is important for providing copyright information and increasing project credibility. In addition, the gray logo helps create a visual contrast with the rest of the site and makes it more discreet.