:sparkles: feat: added Home page to project
This update adds the Home page to the project. The page was created with React and contains the Header component, the Section component and the Footer component.

On the Home page, the Header component displays the project's logo and a simple navigation menu. The Section component displays a title and introductory text. The Footer component displays the project's gray logo and copyright phrase.

Creating the Home page is important to provide a starting point for the project and provide initial information to users. It will be improved over time to meet project needs and improve the user experience.

This update is just the beginning and more enhancements will be added to the Home page in the future.