:lipstick: feat: added initial styling to Home page
This update adds initial styling to the Project Home page.

The initial styling of the Home page is important to provide a coherent and visually pleasing experience for users. With proper application of styles, the page becomes easier to use and more attractive to the target audience.

This update is just the beginning and more improvements will be added to the Home page styling in the future.