:ok_hand: style: color change
This update changes the GRADIENTS_200 and GRADIENTS_100 colors in the theme.js file.

These changes were made to improve the project's visual identity and ensure that the chosen colors are more suitable for the target audience. The new colors were chosen based on color studies and color psychology to ensure they have the desired impact on users.

This update is important to ensure that the project maintains a coherent and attractive visual identity. With the new colors, the project is more modern and attractive to the target audience.