:heavy_plus_sign: build: inclusion of the swiper.js library
This update adds the swiper.js library to the project. The library will be used to create carousels and slides in various parts of the project.

The inclusion of the swiper.js library is important to ensure that the project has a more interactive and modern user experience. With this library, it will be possible to create carousels and slides that will allow users to easily navigate between different contents.