:sparkles: feat: added the Carousel component
This update adds the Carousel component made with Swiper to the project. The component was created to allow users to navigate between different contents in an interactive carousel.

The Carousel component was created using the Swiper library previously added to the project. The carousel is responsive and can be customized in many ways, such as changing the number of slides displayed and the scrolling speed.

This update is just the beginning and more enhancements will be added to the Carousel component in the future to further improve the user experience.