:heavy_plus_sign: build: adding React Router DOM
This update adds the React Router DOM to the project. React Router DOM is a library that lets you manage routes in React apps easily and efficiently. With the React Router DOM, it will be possible to create routes to different pages in the application and ensure that users can easily navigate between them.