:heavy_plus_sign: build: addition of Axios
This update adds Axios to the project. Axios is a library that allows you to make HTTP requests in the browser and in Node.js. With Axios, you can make requests to external APIs and get data to be displayed in the application. Furthermore, Axios is highly configurable and customizable, allowing developers to modify it to suit their specific needs.

Adding Axios also includes initial setup for making requests to an external API.