:sparkles: feat: breakpoints on the slider
This update adds breakpoints to Swiper so that in different screen dimensions it presents itself in different ways so as not to break the layout and make it difficult to see the information present on the cards.

Added 3 different breakpoints for 3 different screen sizes:

- 640 pixels wide or less: Sets one slide at a time and 20 pixels space between them.

- Between 640 and 768 pixels wide: Sets two slides at a time with a 30-pixel gap between them.

- Between 768 and 1024 pixels wide: Sets three slides at a time with a 40-pixel gap between them.

In this way, Swiper adapts according to the screen size, displaying the number of slides and spacing according to the settings defined in each breakpoint.