:lock: feat: added the Auth
This update adds oAuth being responsible for setting an authentication context using createContext, used to store user authentication information.

Through functions, it performs several tasks, such as:

The signIn function is used to authenticate the user, it sends a POST request to the API to login with the user's email and password. If the credentials are valid, the user's token is stored in the browser's local storage and added to the authorization headers in all future requests to the API.

The signOut function removes user information from the browser's local storage and component state.

The signOut function removes user information from the browser's local storage and component state.

The updateProfile function is used to update the user's profile. If the user uploads an avatar file, it is sent to the API to be saved. User information is updated in the browser's API and local storage.

Finally, the useAuth function is a hook that can be used to access the authentication context from other components. It returns the authentication context that was defined in the AuthProvider component.