:lock: feat: typed password display
This update adds the handleKeyDown function to allow triggering the login action when pressing the Enter key after entering the password. In addition, the handleTogglePassword function was implemented to manipulate the variable that controls the display of the password in the password type input field, allowing the user to view the entered password.