:sparkles: feat: user registration
This update carried out several implementations in the system, with emphasis on the connection with the API, which allowed the integration of the system with other platforms.

In addition, the handleSignUp function was developed, whose purpose is to create a new user. It is worth mentioning that, before allowing the creation, the function checks if all the necessary information has been provided by the user and if the password has at least 6 characters. Otherwise, the user is not created.

The handleKeyDown function was also created, being activated when pressing the Enter key after entering the password. When triggered, the function calls the handleSignUp function to create the new user.

To improve the user experience, a Link component was included, which allows navigation between the different pages of the system in an intuitive and simple way.