:lipstick: feat: textual selection
This update disabled the textual selection feature present in the application in order to prevent the user from selecting the text displayed in the interface. This decision was taken considering that there is no need to select text, since all interactions with the system occur through clicks and data entry.