:sparkles: feat: choice of routes
This update added user authentication system integration to the index.jsx file located in the routes directory. With this implementation, it was possible to define that, if a user is authenticated, the Home page will be rendered for him. Otherwise, the Login page will be displayed, offering the option of redirecting to the user registration page, if the user is not yet registered in the application.