:lipstick: feat: positioning change
This update involved a change in the position of the icon responsible for showing or hiding the password in the login form. Previously, the icon position setting was 'bottom: 50%;'. However, due to the change in parent component spacings, this setting has been updated to 'bottom: 43%;'. This measure was implemented to allow the icon to follow the line of text in the form, regardless of the screen size on which the application is being rendered. With this change, it is expected that the user experience will be improved, making interaction with the application easier and more intuitive.