:recycle: refactor: behaviors and icons
This update defined that the application's forms lost the autoComplete functionality. This change was implemented to prevent forms layout from breaking. Also, the 'RiEyeCloseLine' icon has been replaced with 'RiEyeOffLine' to represent the closed eye when showing the hidden password. This change was made to better serve the purpose of hiding the password, providing greater visual clarity for the user. We believe this update will improve the user experience when using the application, making the login process easier and more intuitive.