:bug: fix: data writing and saving error
This update made a correction in storing data in localStorage, as there was an error in writing data that resulted in incorrect insertion, preventing the user from remaining logged in to the application and updating his data. With this correction, the user will be able to remain logged in and have his data updated correctly.