:sparkles: feat: profile page creation
This update implemented the profile page in the project, which was developed with the aim of allowing the user to view their information stored in the application and be able to update them if necessary. The page was created based on previously defined design requirements and included UI elements to improve the user experience. Furthermore, it is foreseen that in a next update of the application, a 'delete account' button will be added on the profile page, which will allow the user to completely remove his account from the application.