:sparkles: feat: default avatar
This update includes the addition of a default in-app avatar, which will be used by users who have not yet entered a profile picture. This avatar consists of a default image stored in the project's image directory and is automatically displayed if the user has not yet selected a custom profile image. This functionality is intended to improve the user experience, maintaining a uniform appearance for the application's interface, even for those who have not yet defined a profile image.