:sparkles: feat: redirects
This update adds a Link element to the Header component that allows redirection to specific parts of the application. In addition, the connection with the API was added to enable the rendering of the logged-in user's avatar. When the user clicks on this avatar, he will be redirected to his profile page. With this, there is an improvement in user navigation in the application, which now has a more practical way of accessing your profile information.