:sparkles: feat: icons
This update added the ability to include icons in its container to the Input component to improve the visual appearance of the application. This will allow developers to further customize the forms and add relevant icons to provide more clarity and ease of use for the user. Additionally, this functionality can improve usability on mobile devices, where screen space is limited and the inclusion of icons can help highlight application functionality.