:sparkles: feat: conditional rendering of icons
This update implements conditional rendering for regular users and administrators. This allows the application to display a visual demonstration of the user who is accessing it, as well as the specific functions available for each type of user. In addition, there was an update to the icon package, changing from 'react-icons/fi' to 'react-icons/ri', thus ensuring consistency and uniform design in the application.