:bug: fix: layout break
This update introduced a fix to the title present on the home page to improve behavior on smaller screens. The issue identified was that the title was breaking off the bounded edges of the screen, which was causing unwanted behavior.

To solve this problem, the 'width' attribute of the title was changed, changing it from '100vh' to '100%'. This modification ensures that the title behaves according to the designer's settings and automatically adjusts to smaller screens without going beyond the screen's borders.