:sparkles: feat: navigation
This update added new functionality to the dish creation page. The 'handleBack' function was implemented, which allows the user to go back to the previous page, which was open before starting the process of creating a new dish.

The 'handleBack' function is an important feature to improve the usability of the application, allowing the user to easily navigate between pages without losing their progress. In addition, this functionality makes it possible to prevent the user from feeling lost or confused when using the application, improving their overall experience.