:sparkles: feat: conditional rendering
This update included connections to the authentication module (auth) in the 'Card' and 'Detail' components of the application. With this implementation, it is possible to identify whether the user who is logged into the application is an administrator or not, and then render the components in different ways, according to the user's access permissions.