:lipstick: feat: icon stylization
This update added 'hover' functionality to the icon present in the 'card' component, which is rendered when the user is logged in as an administrator in the application. When the user hovers the mouse over the icon, a specific visual effect occurs that indicates that the user can click on the icon to perform an action.