:bug: fix: input styling
This update consisted of removing the 'Input' component that was previously used in the section for inserting images of dishes and the creation of a new local 'input' field, since the previous component was not behaving according to the design established by the designer of the interface. Thus, with the implementation of the new field, the error was corrected and the functionality was improved.