:sparkles: feat: autocomplete
This update removed the autocomplete functionality from application inputs via the 'autocomplete=off' attribute. This was done to keep the application's design consistent, as the autocomplete feature was hurting usability by leaving the background of the input white. By removing this functionality, inputs are expected to behave more consistently and pleasantly for users.