:lipstick: feat: styling
This update was responsible for adding styles to the link that contains the user's image, to the text that appears below the project's logo to identify the type of user logged in and to the button that redirects to different pages based on the type of user. These elements have been styled to improve the user's visual experience and make the interface more intuitive.

The intention is to ensure that the appearance of the page is in line with the visual identity of the brand and provide better usability for the user.