:sparkles: feat: routes
This update added a new route in the application that allows editing dishes. This functionality will only be available to users with administrator authorization. With this new route, it will be possible to change important information about the dish, such as the photo, name, category, ingredients, price and details. This way, administrators will be able to make adjustments and updates to the list of dishes offered by the application.