:sparkles: feat: dish editing page
This update added a dish edit page in the project, which allows users with administrator authorization to edit dish information, including photo, name, category, ingredients, price and details. In addition, the possibility of excluding a product from this page has been included. With this functionality, the project gains in flexibility and ease of management, since the editing of dishes can be carried out directly in the system.