:sparkles: feat: dish creation
This update implements the connection with the API and functions were created that allow the creation of new dishes by administrator users. This means that it is now possible to send the necessary information to the API and store it in the database, including the photo, name, category, ingredients, price and details of the new dish. This functionality is exclusive to users with administrator permissions, in order to guarantee the security and quality of the data entered in the system.