:lipstick: feat: image and paragraph
This update triggers changes to the 'object-fit' image fit object format of home 'home' page images.

The format was changed from 'fill' to 'cover' and a 'border-radius' of '50%' was added to better fit the page design. In addition, the layout was adjusted for desktop viewing, where a 'height' of '5rem' and an 'overflow: hidden' were defined so that all paragraphs have the same size and appearance.