:ok_hand: style: text change
This update changed the text that appears below the logo from 'Happy shopping' to 'Bon appétit'. This change was made to improve the adequacy of the text to the type of page, which is a restaurant, and provide a better user experience. The new text will now be at the top of the page, welcoming users.