:bug: fix: sizes
This update has changed the maximum width of the ingredient input. The change was due to layout issues on larger screens, where the input would have excessive whitespace, detracting from the overall appearance of the page. The solution to this problem was to change the 'max-width' property of the element from '13rem' to '8rem', ensuring that it fits better on different screen sizes.