:sparkles: feat: remove dish
This update implemented the option to delete dishes in the product details screen for users with admin authorization. Prior to this update, you had to go to the product's edit page to delete the product. The addition of this functionality makes it easier to administer the application's products, allowing administrators to manage them more efficiently.