:sparkles: feat: search
This update adds the type='search' attribute to the input element of the page in order to conform to current HTML practices and make it more semantic. In addition, this change can also help improve the SEO (Search Engine Optimization) of the site, making it easier to be found by search engines. The type='search' attribute indicates that the field is intended to receive search queries and may allow browsers and other software to provide users with additional search features such as keyword suggestions and autocomplete.