:sparkles: feat: favorites
This update added to the system functions responsible for allowing users to add or remove dishes from their favorites. This means that each dish can be independently added or removed by users. To implement this functionality, new files were created in the system, including routes to create, show, list and delete favorite dishes, the controller to receive information sent by the front-end and respond to requests, the service to process requests and execute the operations of adding and removing bookmarks, and the repository to handle the connection to the database and perform the necessary operations.