:bricks: ci: navigate
This update implements a feature that redirects the user to the home page of the site after creating a new dish. In this way, after finishing the creation process, the user is automatically redirected to the website's home page, instead of remaining on the dish creation page. This change aims to improve the user experience when using the site, making the process more fluid and intuitive.