:lipstick: feat: layout suitability
This update has changed the internal spacing (padding) of components so that they are better suited to screens of different sizes, including small and large screens. A maximum width value for the platters has also been added to prevent them from skewing in different proportions. In addition, there have been changes to the position of elements in the column layout in order to make them more suitable for large screens. These changes are aimed at improving the user experience and ensuring proper presentation of elements on different devices.