:lipstick: feat: buttons
This update implements the behavior of the 'delete' button on the dish details page, as well as its styling. When clicking on the button, the user is asked to confirm that he really wants to delete the item. If confirmed, the item is removed from the database and the page is updated. The button has also been styled to suit the page layout and provide a good user experience.