:sparkles: feat: favorites page
This update adds the 'Favorites' page to the app. This page was created based on the structure of the 'Home' page and the styling was modified to suit the new functionality. The new page displays a list of dishes marked as favorites by the user, using the 'fetchDishes' function to fetch these products from the database and render them on screen.