:bug: fix: footer breaking layout
A behavior problem was detected in the footer on the pages for creating and editing dishes, which affected the integrity of the proposed layout. To fix this problem, a last line called 'footer' has been inserted in the 'grid-template-areas' of these pages. Also, the 'Footer' component has been given the 'grid-area: footer' property to ensure that the footer is in the correct position. In this way, the bug is solved and the layout is working correctly.