:bug: fix: footer breaking layout
A behavior problem was detected in the footer on the dish details page similar to that reported on the dish creation and editing page, which affected the integrity of the proposed layout. To correct this problem, a last line called 'footer' was inserted in the 'grid-model-areas' of the page. In this way, the bug is solved and the layout is working correctly.