:sparkles: feat: itens carts
This update added the 'ItemsCart' component to the project in order to render the dishes added to the user's shopping cart, allowing the user to view and delete them individually. The inclusion of this component aims to improve the user experience when using the platform, allowing them to have more control over what they are adding to the cart and to be able to remove unwanted items in a simple and intuitive way.