:sparkles: feat: payment
This update added the payment component to the project, which will be responsible for managing the payment method selected by the user. It will allow the user to finalize the purchase by sending the paid order to the admin so that he can start preparing and sending the order to the customer.