:lipstick: feat: image
This update added to the project the images that will be rendered in the payment process, including a clock, a QR Code verification item, and a fork and knife image that will be displayed at the end of the transaction. These images are important for the user experience and to make the payment process more intuitive and visually appealing. The clock can indicate the remaining time for the user to complete the purchase, while the QR Code verification item can be used to authenticate the payment. The fork and knife image is displayed at the end of the process as a visual confirmation of the purchase made.