:sparkles: feat: carts
The update added a page called 'Carts' to the project, which will be responsible for rendering the 'ItemsCart' and 'Payment' components. These components were previously created and will be rendered when the user wants to complete the purchase. The 'Carts' page will provide an overview of the items in the cart and allow the user to complete the purchase by selecting a payment method. In addition, the page will also contain relevant information such as the total purchase amount, the number of items and other relevant details.