:sparkles: feat: orders
In this update, the 'Orders' page has been added to the project, with the aim of presenting the purchase history to regular users and allowing administrators to change the status of orders in progress.

For regular users, the page displays information about their purchases, such as order status, quantity of products purchased, which products were purchased, and date of purchase. In addition, you can view which orders are being prepared and wait for the status update.

As for administrator users, the page allows changing the status of orders as they are received. It is also possible to view information about orders, such as quantity and date of completion.