:lipstick: feat: pages and components
This update added stylesheets to the 'ItemsCart' and 'Payment' components, as well as the 'Cart' and 'Orders' pages. Styles were implemented to improve the visual presentation of these elements and pages, making the user experience more pleasant and intuitive. Responsive design techniques were used to ensure that the interface is adaptable to different screen sizes, from mobile devices to desktops.