:lipstick: feat: header
The update carried out in the project promoted the reformulation of the style of the Header component, with the objective of making it more adequate to the new pages available in the project and also more responsive to different types of devices. For this, more semantic classes and separation by modules were created, especially for the desktop version. This update aims to improve the user experience when browsing the site on different devices, allowing them to have a clearer and more organized view of the content.