:sparkles: feat: data entry
In this update, the API connection was added to persist data in the database when the user clicks 'insert' on the page. The handleAddDish function is now asynchronous to ensure the insertion into the database is done correctly and efficiently.

In addition, the price format in Brazilian currency (R$) was added to improve the visualization of prices for the user.