:sparkles: feat: quantity products
In this update, there has been a change to the way the QuantityProducts component is called. Now, it receives two parameters, the 'dish_id' and the 'price', and the entire function of adding the product to the cart has been transferred to this component. This means that when clicking on the add to cart button, the function will be executed within the QuantityProducts component, which is responsible for managing the quantities and prices of the added products.