:lipstick: feat: links and text
In this update, style properties have been added for links in text format. The 'LIGHT_200' color style was applied to the text color of the links, while the 'CAKE_200' color was applied when the mouse cursor is positioned over the link, ie on hover. Such visual properties can help to highlight and draw the user's attention to the links, improving the page's usability and aesthetics.