:sparkles: feat: price and quantity format of products
In this update, the Button component has been removed from the project and will no longer be used. A standardization was made in the rendering of the price in the pt-br format for a better visualization and understanding by the user. In addition, the QuantityProducts component was added in place of the Button component, which will be responsible for displaying the quantity of products selected by the user and allowing the addition of these products to the shopping cart.