:lipstick: feat: fonts
This update made a standardization of the fonts used in the application. The font 'Poppins' was used for the titles and the font 'Roboto' for the rest of the body text. This standardization of fonts contributes to greater visual consistency and improves the readability of texts, making the user experience more pleasant.