:sparkles: feat: password
This update added to the SignUp page the option to show or hide the password entered by the user during the process of creating a new account. This functionality aims to provide more security to the user by allowing him to verify the password before confirming the registration. The option to show or hide the password is triggered by clicking on a corresponding icon next to the password field.