:lipstick: feat: img
This update added images to the '.github' folder to be included in the project's 'README.md' file, in order to demonstrate how the application looks when run for standard users and for administrator users. The images serve as illustrations to show the differences in layout and features available for each type of user. This addition can make it easier to understand how the application works and help new developers get familiar with the project.