:iphone: placement and size
This update was focused on making changes to the application's layout and interface so that they are compatible and adjustable on mobile devices. The goal is to improve the user experience and make the application more accessible. The changes included resizing the placement of components and the size of the fonts used. Such changes are important to ensure that the application is accessible on different mobile devices, such as smartphones and tablets.