:lipstick: feat: delete
In this update, the delete dish button has been restyled, adding a hover effect to improve the user experience.

This delete button style improvement aims to make the delete action more intuitive and standout for the user, avoiding confusion and misunderstandings. By adding the hover effect, the button visually highlights when the cursor is moved over it, providing clear visual feedback to the user that they can interact with the button to perform the delete.